refactor: Move the useCurrentEditInformations hook to Hooks folder
